Review comments incorporation for separate nuget install
