+++ tools/pipelines-tasks/azure-pipelines/build-vsix.yml
@@ -17,8 +17,8 @@ variables:
   tasksRoot: 'tools/pipelines-tasks'
   buildOutRoot: '$(Build.ArtifactStagingDirectory)/buildOutput'
   # Version number
-  major: '2'
-  minor: '0'
+  major: '1'
+  minor: '2'
   patch: $[counter(variables['minor'], 0)]
   # Upload results of Semmle analysis
   LGTM.UploadSnapshot: true
