+++ tools/pipelines-tasks/common/.npmignore
@@ -1,0 +1,2 @@
+installAppAttachFramework.*
+NugetInstall.ps1
