+++ tools/pipelines-tasks/vss-extension.json
@@ -2,7 +2,7 @@
   "manifestVersion":  1,
   "id": "msix-ci-automation-task",
   "name":  "MSIX Packaging",
-  "version":  "2.0.0",
+  "version":  "1.2.0",
   "publisher":  "MSIX",
   "targets":  [
     {
